#Task3.7: FavoritesList component
